Uncertainty UI Enhancements (#715)
* Added confidence_interval as an optional parameter for NeuralProphet().

* Added confidence_interval as an optional parameter for NeuralProphet().

* Added uncertainty_estimation param and renamed confidence_interval param to interval_width.

* Modified uncertainty test cases so it passes with uncertainty_estimation and interval_width params and speed up the uncertainty tests.

* Renamed NeuralProphet param interval_width to prediction_interval.

* Moved the uncertainty estimation tests from test_integration.py to new file test_uncertainty.py.

* Renamed uncertainty_estimation param to uncertainty_method in forecaster.py.

* Changed uncertainty_method value from 'conformal_prediction' to 'auto' in test_uncertainty_estimation_prediction_interval.

* Removed 'conformal_prediction' from ValueError and changed if-statement to  for configure.py.

* Changed the interval_width param to prediction_interval for tutorials/feature-use/uncertainty_estimation_peyton_manning.ipynb.

* Clarified assertion error statement for uncertainty_method param in configure.py and added mini-example for uncertainty_method in the NeuralProphet comment section in forecaster.py.

* Modified the uncertainty_method assert statement in configure.py and break it up to make it less lengthy.

* Removed the 'conformal_prediction' option for the uncertainty_method param comment in forecaster.py.
